Moved the LightLightness model to a dedicated file
