+++ bluetooth_mesh/models/light/__init__.py
@@ -1,0 +1,3 @@
+# pylint: disable=W0614
+
+from .light_lightness import *
