+++ bluetooth_mesh/models/models.py
@@ -2067,286 +2067,6 @@ class GenericLevelClient(Model):
         }
 
 
-class LightLightnessServer(Model):
-    MODEL_ID = (None, 0x1300)
-    OPCODES = {
-        LightLightnessOpcode.LIGHT_LIGHTNESS_GET,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_SET,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_SET_UNACKNOWLEDGED,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class LightLightnessSetupServer(Model):
-    MODEL_ID = (None, 0x1301)
-    OPCODES = {
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_DEFAULT_SET,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_DEFAULT_SET_UNACKNOWLEDGED,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET_UNACKNOWLEDGED,
-    }
-    SUBSCRIBE = True
-
-
-class LightLightnessClient(Model):
-    MODEL_ID = (None, 0x1302)
-    OPCODES = {
-        LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_lightness_range_unack(
-        self,
-        destination: int,
-        app_index: int,
-        min_lightness: int,
-        max_lightness: int,
-        *,
-        retransmissions: int = 6,
-        send_interval: float = 0.075,
-    ) -> None:
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET_UNACKNOWLEDGED,
-                params=dict(
-                    range_min=min_lightness,
-                    range_max=max_lightness,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request, retransmissions=retransmissions, send_interval=send_interval
-        )
-
-    async def set_lightness_range(
-        self,
-        destination: int,
-        app_index: int,
-        min_lightness: int,
-        max_lightness: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Optional[Any]:
-        request = partial(
-            self.send_app,
-            destination=destination,
-            app_index=app_index,
-            opcode=LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET,
-            params=dict(
-                range_min=min_lightness, range_max=max_lightness, tid=self.tid()
-            ),
-        )
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS
-
-        status = self.expect_app(
-            source=destination,
-            app_index=0,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        result = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 2.0,
-        )
-
-        return result[status_opcode.name.lower()]
-
-    async def get_lightness_range(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.075,
-        timeout: Optional[float] = None,
-    ) -> None:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def get_lightness(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set_lightness_unack(
-        self,
-        destination: int,
-        app_index: int,
-        lightness: int,
-        transition_time: float = 0,
-        *,
-        delay: float = 0.5,
-        retransmissions: int = 6,
-        send_interval: float = 0.075,
-    ) -> None:
-        tid = self.tid()
-        remaining_delay = delay
-
-        async def request():
-            nonlocal remaining_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_SET_UNACKNOWLEDGED,
-                params=dict(
-                    lightness=lightness,
-                    delay=remaining_delay,
-                    tid=tid,
-                    transition_time=transition_time,
-                ),
-            )
-            remaining_delay = max(0.0, remaining_delay - send_interval)
-
-            return await ret
-
-        await self.repeat(
-            request, retransmissions=retransmissions, send_interval=send_interval
-        )
-
-    async def set_lightness(
-        self,
-        nodes: Sequence[int],
-        lightness: int,
-        app_index: int,
-        transition_time: float = 0,
-        *,
-        delay: float = 0.5,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        print("transition_time2=%f" % transition_time)
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_SET,
-                params=dict(
-                    lightness=lightness,
-                    tid=self.tid(),
-                    delay=0,
-                    transition_time=transition_time),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
 class SensorClient(Model):
     MODEL_ID = (None, 0x1102)
     OPCODES = {
